backend/requirements.txt
flask>=3.0
flask-cors>=4.0
opencv-python-headless>=4.8
numpy>=1.24
scipy>=1.11
torch>=2.0
pytorch-lightning>=2.0
transformers>=4.30
torchmetrics>=1.0
pillow>=10.0
